added mss calculation, using any IPASIR sat solver; changed long long to int, since an int is 32 bits on modern computers, and 32 bits suffices for the benchmarks; ajusted some data structures of leximaxIST::Encoder

+++ no-clause-set/include/leximaxIST_parsing_utils.h
@@ -134,9 +134,9 @@ namespace leximaxIST {
     return rstr; }
 
     template<class B>
-    static long long parseInt(B& in) {
-        long long    val = 0;
-        bool    neg = false;
+    static int parseInt(B& in) {
+        int    val = 0;
+        bool   neg = false;
         skipWhitespace(in);
         if      (*in == '-') neg = true, ++in;
         else if (*in == '+') ++in;
